samples/web-app-managed-identity/python/src/requirements.txt
Flask==3.1.0
azure-identity==1.12.0
azure-storage-blob==12.26.0
azure-core==1.35.0
gunicorn==20.1.0
python-dotenv==1.1.1